backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
httpx>=0.24.0,<0.25.0
openai==1.3.0
psycopg2-binary==2.9.10
